Change paths to css files + dumbify App component
