+++ src/components/Card/Card.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Card = ({ itemId = 0, ...props }) => (
+  <article>i am a pretty stupid card!</article>
+);
+
+export default Card;
