+++ src/components/Input/Input.scss
@@ -1,0 +1,19 @@
+input {
+    border-radius: var(--border-rounded);
+    padding: 0.5em;
+    font-weight: 400;
+}
+
+.default {
+    background: var(--color-default);
+}
+
+.primary {
+    background: var(--color-primary);
+    color: var(--color-light);
+}
+
+.secondary {
+    background: var(--color-secondary);
+    color: var(--color-light);
+}
