+++ src/components/index.js
@@ -1,2 +1,4 @@
-export Input from './Input/Input';
-export Button from './Button/Button';
+export { default as Input } from './Input/Input';
+export { default as Button } from './Button/Button';
+export { default as Card } from './Card/Card';
+export { default as App } from './App/App';
